# Conflicts:
#	code/packages/rust/aarch64-backend/CHANGELOG.md
#	code/packages/rust/aarch64-backend/Cargo.toml
#	code/packages/rust/aarch64-backend/src/lib.rs
#	code/packages/rust/iir-to-cil-bytecode/CHANGELOG.md
#	code/packages/rust/iir-to-cil-bytecode/Cargo.toml
#	code/packages/rust/iir-to-jvm-class-file/CHANGELOG.md
#	code/packages/rust/iir-to-jvm-class-file/Cargo.toml
#	code/packages/rust/iir-to-jvm-class-file/src/lower.rs
#	code/packages/rust/iir-to-jvm-class-file/src/validate.rs
#	code/packages/rust/iir-to-llvm/CHANGELOG.md
#	code/packages/rust/iir-to-llvm/Cargo.toml
#	code/packages/rust/iir-to-wasm/CHANGELOG.md
#	code/packages/rust/iir-to-wasm/Cargo.toml
#	code/packages/rust/iir-to-wasm/src/lower.rs
#	code/packages/rust/iir-to-wasm/src/validate.rs
#	code/packages/rust/lang-aot/CHANGELOG.md
#	code/packages/rust/lang-aot/Cargo.toml
#	code/packages/rust/lang-aot/tests/lang_matrix.rs
#	code/packages/rust/x86_64-backend/CHANGELOG.md
#	code/packages/rust/x86_64-backend/Cargo.toml
#	code/packages/rust/x86_64-backend/src/lib.rs
#	code/specs/SIR24-semantic-ir-to-c.md

+++ code/packages/python/audio-device-sink/src/audio_device_sink/__init__.py
@@ -45,7 +45,7 @@ def _integer_sample_rate(value: Any) -> int:
         converted = float(value)
         if not isfinite(converted) or converted != round(converted):
             raise ValueError("sample_rate_hz must be an integer-valued rate")
-        rate = int(round(converted))
+        rate = round(converted)
     else:
         raise TypeError("sample_rate_hz must be an integer")
 
@@ -127,7 +127,7 @@ def play_pcm_buffer(buffer: Any) -> PlaybackReport:
 
 __version__ = "0.1.0"
 
-__all__ = [
+__all__ = [  # noqa: RUF022 - preserve the package's stable export order
     "AudioDeviceError",
     "MAX_BLOCKING_DURATION_SECONDS",
     "MAX_SAMPLE_RATE_HZ",
